+++ src/configuration.ts
@@ -84,7 +84,10 @@ const configuration = {
     URL: process.env.RAWG_API_URL || "https://api.rawg.io/api",
     KEY: process.env.RAWG_API_KEY || "",
     CACHE_DAYS: Number(process.env.RAWG_API_CACHE_DAYS) || 30,
-    EXCLUDE_STORES: process.env.RAWG_API_EXCLUDE_STORES || "9",
+    INCLUDED_STORES: parseList(
+      process.env.RAWG_API_INCLUDED_STORES,
+      globals.DEFAULT_INCLUDED_RAWG_STORES,
+    ),
   } as const,
   USERS: {
     REQUIRE_EMAIL: parseBooleanEnvVariable(process.env.USERS_REQUIRE_EMAIL),